make schemas more modular to support response validation and documentation
