+++ src/api/v1/metadata/schemas/index.js
@@ -1,0 +1,4 @@
+import { metadataParamSchema } from './params.js'
+import { responseSchemas } from './responses.js'
+
+export { metadataParamSchema, responseSchemas }
