+++ test/mocks/metadata.js
@@ -1,17 +1,6 @@
 export const mockFormattedMetadata =
 [
   {
-    raw: {
-      fileId: '9fcaabe5-77ec-44db-8356-3a6e8dc51b13',
-      filename: 'dragon-b.jpeg',
-      contentType: 'image/jpeg',
-      fileStatus: 'complete',
-      contentLength: 11264,
-      checksumSha256: 'bng5jOVC6TxEgwTUlX4DikFtDEYEc8vQTsOP0ZAv21c=',
-      detectedContentType: 'image/jpeg',
-      s3Key: '3b0b2a02-a669-44ba-9b78-bd5cb8460253/9fcaabe5-77ec-44db-8356-3a6e8dc51b13',
-      s3Bucket: 'cdp-example-node-frontend'
-    },
     metadata: {
       sbi: '105000000',
       crn: '1050000000',
@@ -30,24 +19,9 @@ export const mockFormattedMetadata =
       filename: 'dragon-b.jpeg',
       contentType: 'image/jpeg',
       fileStatus: 'complete'
-    },
-    s3: {
-      key: '3b0b2a02-a669-44ba-9b78-bd5cb8460253/9fcaabe5-77ec-44db-8356-3a6e8dc51b13',
-      bucket: 'cdp-example-node-frontend'
     }
   },
   {
-    raw: {
-      fileId: '3f90b889-eac7-4e98-975f-93fcef5b8554',
-      filename: 'Health and Safety Assessment Certificate.pdf',
-      contentType: 'application/pdf',
-      fileStatus: 'complete',
-      contentLength: 115307,
-      checksumSha256: 'ZbILFUsbS2Pio0Sv2ifwyg+SSQsnzVF1h6fQzAiBt4Q=',
-      detectedContentType: 'application/pdf',
-      s3Key: 'scanned/8ea63b47-f5ac-410e-8b1c-3ae522b0a96c/3f90b889-eac7-4e98-975f-93fcef5b8554',
-      s3Bucket: 'fcp-sfd-object-processor-bucket'
-    },
     metadata: {
       sbi: '105000000',
       crn: '1050000000',
@@ -66,10 +40,27 @@ export const mockFormattedMetadata =
       filename: 'Health and Safety Assessment Certificate.pdf',
       contentType: 'application/pdf',
       fileStatus: 'complete'
-    },
-    s3: {
-      key: 'scanned/8ea63b47-f5ac-410e-8b1c-3ae522b0a96c/3f90b889-eac7-4e98-975f-93fcef5b8554',
-      bucket: 'fcp-sfd-object-processor-bucket'
     }
   }
 ]
+
+export const mockRawData = {
+  raw: {
+    fileId: '9fcaabe5-77ec-44db-8356-3a6e8dc51b13',
+    filename: 'dragon-b.jpeg',
+    contentType: 'image/jpeg',
+    fileStatus: 'complete',
+    contentLength: 11264,
+    checksumSha256: 'bng5jOVC6TxEgwTUlX4DikFtDEYEc8vQTsOP0ZAv21c=',
+    detectedContentType: 'image/jpeg',
+    s3Key: '3b0b2a02-a669-44ba-9b78-bd5cb8460253/9fcaabe5-77ec-44db-8356-3a6e8dc51b13',
+    s3Bucket: 'cdp-example-node-frontend'
+  }
+}
+
+export const mockS3Data = {
+  s3: {
+    key: '3b0b2a02-a669-44ba-9b78-bd5cb8460253/9fcaabe5-77ec-44db-8356-3a6e8dc51b13',
+    bucket: 'cdp-example-node-frontend'
+  }
+}
